validate float fixet.

+++ database/migrations/2022_01_14_192642_create_products_table.php
@@ -16,7 +16,7 @@ class CreateProductsTable extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('unit');
+            $table->float('unit');
             $table->float('unit_price');
             $table->string('slug');
             $table->timestamps();
